Rebase and resolve remaining imports
