+++ tests-e2e/appstudio/deploymenttargetclaim_binder_test.go
@@ -1,4 +1,4 @@
-package rhtap
+package appstudio
 
 import (
 	"context"
