+++ tests-e2e/core/gitopsdeployment_syncrun_test.go
@@ -22,6 +22,12 @@ import (
 )
 
 var _ = Describe("GitOpsDeploymentSyncRun E2E tests", func() {
+
+	const (
+		name    = "my-gitops-depl"
+		repoURL = "https://github.com/redhat-appstudio/managed-gitops"
+	)
+
 	Context("Create, update and delete GitOpsDeploymentSyncRun", func() {
 
 		var (
@@ -48,7 +54,7 @@ var _ = Describe("GitOpsDeploymentSyncRun E2E tests", func() {
 			Expect(err).To(Succeed())
 
 			By("create a GitOpsDeployment with 'Manual' sync policy")
-			gitOpsDeploymentResource = BuildGitOpsDeploymentResource(name,
+			gitOpsDeploymentResource = gitopsDeplFixture.BuildGitOpsDeploymentResource(name,
 				repoURL, "resources/test-data/sample-gitops-repository/environments/overlays/dev",
 				managedgitopsv1alpha1.GitOpsDeploymentSpecType_Manual)
 			gitOpsDeploymentResource.Spec.Destination.Environment = ""
@@ -76,7 +82,7 @@ var _ = Describe("GitOpsDeploymentSyncRun E2E tests", func() {
 		It("creating a new GitOpsDeploymentSyncRun should sync an Argo CD Application", func() {
 
 			By("create a GitOpsDeploymentSyncRun")
-			syncRunCR := BuildGitOpsDeploymentSyncRunResource("test-syncrun", fixture.GitOpsServiceE2ENamespace, gitOpsDeploymentResource.Name, "main")
+			syncRunCR := syncRunFixture.BuildGitOpsDeploymentSyncRunResource("test-syncrun", fixture.GitOpsServiceE2ENamespace, gitOpsDeploymentResource.Name, "main")
 
 			err := k8sClient.Create(ctx, &syncRunCR)
 			Expect(err).To(BeNil())
@@ -100,7 +106,7 @@ var _ = Describe("GitOpsDeploymentSyncRun E2E tests", func() {
 		It("updating the spec of an existing GitOpsDeploymentSyncRun CR should not trigger a new sync", func() {
 
 			By("create a GitOpsDeploymentSyncRun")
-			syncRunCR := BuildGitOpsDeploymentSyncRunResource("test-syncrun", fixture.GitOpsServiceE2ENamespace, gitOpsDeploymentResource.Name, "main")
+			syncRunCR := syncRunFixture.BuildGitOpsDeploymentSyncRunResource("test-syncrun", fixture.GitOpsServiceE2ENamespace, gitOpsDeploymentResource.Name, "main")
 
 			err := k8sClient.Create(ctx, &syncRunCR)
 			Expect(err).To(BeNil())
@@ -152,7 +158,7 @@ var _ = Describe("GitOpsDeploymentSyncRun E2E tests", func() {
 
 		It("applying the same CR with no changes should not trigger a new sync", func() {
 			By("create a GitOpsDeploymentSyncRun")
-			syncRunCR := BuildGitOpsDeploymentSyncRunResource("test-syncrun", fixture.GitOpsServiceE2ENamespace, gitOpsDeploymentResource.Name, "main")
+			syncRunCR := syncRunFixture.BuildGitOpsDeploymentSyncRunResource("test-syncrun", fixture.GitOpsServiceE2ENamespace, gitOpsDeploymentResource.Name, "main")
 
 			err := k8sClient.Create(ctx, &syncRunCR)
 			Expect(err).To(BeNil())
@@ -199,7 +205,7 @@ var _ = Describe("GitOpsDeploymentSyncRun E2E tests", func() {
 		})
 
 		It("deleting the GitOpsDeploymentSyncRun should terminate a running Sync operation", func() {
-			gitOpsDeploymentResource = BuildGitOpsDeploymentResource("test-deply-with-presync",
+			gitOpsDeploymentResource = gitopsDeplFixture.BuildGitOpsDeploymentResource("test-deply-with-presync",
 				"https://github.com/managed-gitops-test-data/deployment-presync-hook", "guestbook",
 				managedgitopsv1alpha1.GitOpsDeploymentSpecType_Manual)
 			gitOpsDeploymentResource.Spec.Destination.Environment = ""
@@ -224,7 +230,7 @@ var _ = Describe("GitOpsDeploymentSyncRun E2E tests", func() {
 			addCustomHealthCheckForDeployment(ctx, k8sClient, argocdCR)
 
 			By("create a GitOpsDeploymentSyncRun")
-			syncRunCR := BuildGitOpsDeploymentSyncRunResource("test-syncrun", fixture.GitOpsServiceE2ENamespace, gitOpsDeploymentResource.Name, "main")
+			syncRunCR := syncRunFixture.BuildGitOpsDeploymentSyncRunResource("test-syncrun", fixture.GitOpsServiceE2ENamespace, gitOpsDeploymentResource.Name, "main")
 
 			err = k8sClient.Create(ctx, &syncRunCR)
 			Expect(err).To(BeNil())
@@ -270,7 +276,7 @@ var _ = Describe("GitOpsDeploymentSyncRun E2E tests", func() {
 		It("deleting the GitOpsDeploymentSyncRun CR should not terminate if no Sync operation is in progress", func() {
 
 			By("create a GitOpsDeploymentSyncRun")
-			syncRunCR := BuildGitOpsDeploymentSyncRunResource("test-syncrun", fixture.GitOpsServiceE2ENamespace, gitOpsDeploymentResource.Name, "main")
+			syncRunCR := syncRunFixture.BuildGitOpsDeploymentSyncRunResource("test-syncrun", fixture.GitOpsServiceE2ENamespace, gitOpsDeploymentResource.Name, "main")
 
 			err := k8sClient.Create(ctx, &syncRunCR)
 			Expect(err).To(BeNil())
@@ -318,7 +324,7 @@ var _ = Describe("GitOpsDeploymentSyncRun E2E tests", func() {
 
 		It("should sync if the previous sync operation is terminated", func() {
 
-			gitOpsDeploymentResource = BuildGitOpsDeploymentResource("test-deply-with-presync",
+			gitOpsDeploymentResource = gitopsDeplFixture.BuildGitOpsDeploymentResource("test-deply-with-presync",
 				"https://github.com/managed-gitops-test-data/deployment-presync-hook", "guestbook",
 				managedgitopsv1alpha1.GitOpsDeploymentSpecType_Manual)
 			gitOpsDeploymentResource.Spec.Destination.Environment = ""
@@ -343,7 +349,7 @@ var _ = Describe("GitOpsDeploymentSyncRun E2E tests", func() {
 			addCustomHealthCheckForDeployment(ctx, k8sClient, argocdCR)
 
 			By("create a GitOpsDeploymentSyncRun")
-			syncRunCR := BuildGitOpsDeploymentSyncRunResource("test-syncrun", fixture.GitOpsServiceE2ENamespace, gitOpsDeploymentResource.Name, "main")
+			syncRunCR := syncRunFixture.BuildGitOpsDeploymentSyncRunResource("test-syncrun", fixture.GitOpsServiceE2ENamespace, gitOpsDeploymentResource.Name, "main")
 
 			err = k8sClient.Create(ctx, &syncRunCR)
 			Expect(err).To(BeNil())
@@ -375,7 +381,7 @@ var _ = Describe("GitOpsDeploymentSyncRun E2E tests", func() {
 			Expect(err).To(BeNil())
 
 			By("create a new GitOpsDeploymentSyncRun and ensure that the sync status hasn't changed")
-			newSyncRunCR := BuildGitOpsDeploymentSyncRunResource("test-syncrun-1", fixture.GitOpsServiceE2ENamespace, gitOpsDeploymentResource.Name, "main")
+			newSyncRunCR := syncRunFixture.BuildGitOpsDeploymentSyncRunResource("test-syncrun-1", fixture.GitOpsServiceE2ENamespace, gitOpsDeploymentResource.Name, "main")
 
 			err = k8sClient.Create(ctx, &newSyncRunCR)
 			Expect(err).To(BeNil())
@@ -396,7 +402,7 @@ var _ = Describe("GitOpsDeploymentSyncRun E2E tests", func() {
 		})
 
 		It("should handle multiple GitOpsDeploymentSyncRun CRs sequentially", func() {
-			gitOpsDeploymentResource = BuildGitOpsDeploymentResource("test-deply",
+			gitOpsDeploymentResource = gitopsDeplFixture.BuildGitOpsDeploymentResource("test-deply",
 				"https://github.com/managed-gitops-test-data/deployment-presync-hook", "guestbook-without-hook",
 				managedgitopsv1alpha1.GitOpsDeploymentSpecType_Manual)
 			gitOpsDeploymentResource.Spec.Destination.Environment = ""
@@ -418,7 +424,7 @@ var _ = Describe("GitOpsDeploymentSyncRun E2E tests", func() {
 			}
 
 			for _, cr := range syncRunCRs {
-				syncRunCR := BuildGitOpsDeploymentSyncRunResource(cr.name, fixture.GitOpsServiceE2ENamespace, gitOpsDeploymentResource.Name, cr.ref)
+				syncRunCR := syncRunFixture.BuildGitOpsDeploymentSyncRunResource(cr.name, fixture.GitOpsServiceE2ENamespace, gitOpsDeploymentResource.Name, cr.ref)
 
 				err := k8sClient.Create(ctx, &syncRunCR)
 				Expect(err).To(BeNil())
@@ -470,7 +476,7 @@ var _ = Describe("GitOpsDeploymentSyncRun E2E tests", func() {
 			}
 
 			By("create a GitOpsDeploymentSyncRun")
-			syncRunCR := BuildGitOpsDeploymentSyncRunResource("test-syncrun", fixture.GitOpsServiceE2ENamespace, gitOpsDeploymentResource.Name, "main")
+			syncRunCR := syncRunFixture.BuildGitOpsDeploymentSyncRunResource("test-syncrun", fixture.GitOpsServiceE2ENamespace, gitOpsDeploymentResource.Name, "main")
 
 			err := k8sClient.Create(ctx, &syncRunCR)
 			Expect(err).To(BeNil())
